a red dot will be shown to show which tab user is in
